+++ App/Sagas/likesPost/firebase/indnex.spec.js
@@ -2,12 +2,13 @@
 import test from 'ava-spec';
 import {call} from 'redux-saga/effects';
 
-import {likeWithKeyMock} from '../../../Tests/mock/';
-import { getLikeWithChannelId } from '../../Services';
-import {likesPostToFirebase, increaseOnFirebase} from './firebase';
+import {likeWithKeyMock} from '../../../../Tests/mock/';
+import { getLikeWithChannelId } from '../../../Services';
+import {increase, increaseOnFirebase} from './increase';
+import {_new} from './new';
 
 test.serial.group('Normal', () => {
-  const generator = likesPostToFirebase.increase('channelId', 1, 'uid');
+  const generator = increase('channelId', 1, 'uid');
 
   test('could check if the target like is already on server', (t) => {
     t.deepEqual(
@@ -31,7 +32,7 @@ test.serial.group('Normal', () => {
 });
 
 test.serial.group('Abnormal', () => {
-  const generator = likesPostToFirebase.increase('channelId', 1, 'uid');
+  const generator = increase('channelId', 1, 'uid');
   generator.next();
 
   test('could push new like to Firebase', (t) => {
@@ -41,13 +42,13 @@ test.serial.group('Abnormal', () => {
         message: '',
         snapshot: null,
       }).value,
-      call(likesPostToFirebase.likesNew, 'channelId', 1, 'uid'),
+      call(_new, 'channelId', 1, 'uid'),
     );
   });
 });
 
 test.serial.group('Abnormal', () => {
-  const generator = likesPostToFirebase.increase('channelId', 1, 'uid');
+  const generator = increase('channelId', 1, 'uid');
   generator.next();
 
   test("could return response when it couldn't get like", (t) => {
